right before sync with upstream

+++ app/models/assessment.rb
@@ -225,8 +225,11 @@ class Assessment < ApplicationRecord
     config = config_source.gsub("module #{source_config_module_name}",
                                 "module #{config_module_name}")
 
-    # backup old config
-    File.rename(config_file_path, config_file_path.sub_ext(".rb.bak"))
+    # backup old config; but only if exists -tue.
+    if(File.exist?(config_file_path))
+      File.rename(config_file_path, config_file_path.sub_ext(".rb.bak"))
+    end
+
 
     # write to config_file_path
     File.open(config_file_path, "w") { |f| f.write config }
